refactor - CarGame 클래스 삭제 및 CarController, CarMain 클래스 수정(CarGame의 gameStart 메소드를 CarClass에서 구현)
