+++ src/main/java/step3/CarMain.java
@@ -6,12 +6,12 @@ public class CarMain {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        CarGame carGame = new CarGame();
+        CarController carController = new CarController();
 
         int numberOfCars = inputView.numberOfCars();
         int tryCounts = inputView.tryCount();
 
-        List<CarList> carListResult = carGame.gameStart(numberOfCars, tryCounts);
+        List<CarList> carListResult = carController.gameStart(numberOfCars, tryCounts);
 
         outputView.resultMention();
         outputView.displayCarListResult(carListResult);
